feat: Add interview and evaluation management templates
- Created templates for candidate interviews including index, show, new, and edit views.
- Developed recruiter evaluation templates for listing, creating, editing, and showing evaluations.
- Implemented a schema update script for database management.

+++ src/Entity/Evaluation_scores.php
@@ -11,6 +11,7 @@ class Evaluation_scores
 {
 
     #[ORM\Id]
+    #[ORM\GeneratedValue]
     #[ORM\Column(type: "integer")]
     private int $score_id;
 
@@ -22,8 +23,8 @@ class Evaluation_scores
     #[ORM\JoinColumn(name: 'criteria_id', referencedColumnName: 'criteria_id', onDelete: 'CASCADE')]
     private Evaluation_criteria $criteria_id;
 
-    #[ORM\Column(type: "integer")]
-    private int $score;
+    #[ORM\Column(type: "float")]
+    private float $score;
 
     #[ORM\Column(type: "text")]
     private string $comments;
